Website/requirements.txt
Django
pandas
efficientnet_pytorch
django-embed-video
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.6.0
Unipath==1.0
python-decouple==3
Pillow==4.1.1
Markdown==2.6.6
bleach==1.4.3
psycopg2
whitenoise==3.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.5.1+cpu 
torchvision==0.6.1+cpu 